♻️ ドロップゾーンの処理をリファクタリング
