+++ src/pages/admins/spots/post/index.tsx
@@ -13,9 +13,6 @@ import { Spot } from 'src/types/spot';
 import { Prefectures } from 'src/types/prefectures';
 import { Systems } from 'src/types/systems';
 import { Spinner } from '@chakra-ui/react';
-import { Group, Text, useMantineTheme, MantineTheme } from '@mantine/core';
-import { Upload, Camera, X, Icon as TablerIcon } from 'tabler-icons-react';
-import { Dropzone, DropzoneStatus, MIME_TYPES } from '@mantine/dropzone';
 import { DEFAULT_SPOTS_BUCKET } from 'src/libs/regular';
 import SpotImage from 'src/components/Spot/SpotImage';
 import { getPrefecturesCreatedAt } from 'src/hooks/usePrefecturesCreatedAtSelect';
@@ -66,7 +63,6 @@ const SpotsPost: NextPage = () => {
   const [error, setError] = useState(false);
 
   const user = supabase.auth.user();
-  const theme = useMantineTheme();
 
   useEffect(() => {
     setSession(supabase.auth.session());
@@ -101,7 +97,7 @@ const SpotsPost: NextPage = () => {
         throw uploadError;
       }
 
-      console.log('ユーザー', user?.id);
+      // console.log('ユーザー', user?.id);
 
       const { error: updateError } = await supabase.from('spots').insert({
         admin_id: user?.id,
@@ -148,57 +144,6 @@ const SpotsPost: NextPage = () => {
       setLoading(false);
     }
   }, []);
-
-  function getIconColor(status: DropzoneStatus, theme: MantineTheme) {
-    return status.accepted
-      ? theme.colors[theme.primaryColor][theme.colorScheme === 'dark' ? 4 : 6]
-      : status.rejected
-      ? theme.colors.red[theme.colorScheme === 'dark' ? 4 : 6]
-      : theme.colorScheme === 'dark'
-      ? theme.colors.dark[0]
-      : theme.colors.gray[7];
-  }
-
-  function ImageUploadIcon({
-    status,
-    ...props
-  }: React.ComponentProps<TablerIcon> & { status: DropzoneStatus }) {
-    if (status.accepted) {
-      return <Upload {...props} />;
-    }
-
-    if (status.rejected) {
-      return <X {...props} />;
-    }
-
-    return <Camera {...props} />;
-  }
-
-  const dropzoneChildren = (status: DropzoneStatus, theme: MantineTheme) => (
-    <Group
-      position='center'
-      spacing='sm'
-      direction='column'
-      style={{ minHeight: 120, pointerEvents: 'none' }}
-    >
-      <div className='flex border-2 border-red-600 rounded-lg p-2 mt-6'>
-        <ImageUploadIcon
-          status={status}
-          style={{ color: getIconColor(status, theme) }}
-          size={20}
-          color='red'
-        />
-        <Text size='sm' color='red' weight={700} inline mt={3} ml={3}>
-          画像を選択する
-        </Text>
-      </div>
-      <div>
-        <Text size='sm' color='blue' weight={700} inline>
-          またはドラッグ&ドロップ
-        </Text>
-      </div>
-    </Group>
-  );
 
   const fetchPrefecturesCreatedAt = useCallback(async () => {
     try {
@@ -284,6 +229,8 @@ const SpotsPost: NextPage = () => {
     // ) {
     //   toast.error('入力されていない項目があります', {});
     // } else {
+    console.log('イメージ', spotPost.image_url);
+
     const { data, error } = await supabase
       .from('spots')
       .update({
@@ -358,14 +305,14 @@ const SpotsPost: NextPage = () => {
               <SpotImage url={spotImage} size={60} />
             ) : (
               <div className='flex flex-wrap gap-2 mt-5 sm:gap-6'>
-                <div className='bg-white w-16 h-16'>
+                {/* <div className='bg-white w-16 h-16'>
                   <img
                     src='/icons/camera-icon.png'
                     alt='カメラアイコン'
                     className='m-auto mt-4 w-8 h-8'
                   />
                 </div>
-                {/* <div className='bg-white w-16 h-16'>
+                <div className='bg-white w-16 h-16'>
                   <img
                     src='/icons/camera-icon.png'
                     alt='カメラアイコン'
@@ -396,16 +343,6 @@ const SpotsPost: NextPage = () => {
               </div>
             )}
             <SpotUploadButton onUpload={uploadSpots} loading={uploading} />
-            <div className='mt-5'>
-              <Dropzone
-                onDrop={(files) => console.log('accepted files', files)}
-                onReject={(files) => console.log('rejected files', files)}
-                maxSize={3 * 1024 ** 2}
-                accept={[MIME_TYPES.png, MIME_TYPES.jpeg, MIME_TYPES.svg, MIME_TYPES.gif]}
-              >
-                {(status) => dropzoneChildren(status, theme)}
-              </Dropzone>
-            </div>
             {/* スポット情報 */}
             <div>
               <h2 className='mt-10 '>スポット情報</h2>
